% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{.display_results}
\alias{.display_results}
\title{Internal function to format and display results}
\usage{
.display_results(x, dVec.range = NULL)
}
\description{
Internal function to format and display results
}
\keyword{internal}
