% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle_average.R
\name{GenSamplingPoints}
\alias{GenSamplingPoints}
\title{Generate sampling points on a circle (edge), disk, or donut}
\usage{
GenSamplingPoints(
  center,
  radius,
  numpts,
  dist.metric = "Euclidean",
  cType = "edge",
  dist_unit = NULL
)
}
\arguments{
\item{center}{Numeric vector c(x, y)}

\item{radius}{Distance radius}

\item{numpts}{Number of points around the circle}

\item{dist.metric}{"Euclidean" or "Geodesic"}

\item{cType}{"edge", "disk", or "donut"}

\item{dist_unit}{Step size for disk/donut radii}
}
\value{
Matrix of (x, y) sampling coordinates
}
\description{
Generate sampling points on a circle (edge), disk, or donut
}
\keyword{internal}
