% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle_average.R
\name{GenSamplingPointsBuffer}
\alias{GenSamplingPointsBuffer}
\title{Generate sampling points by buffering a polygon (sf-based)}
\usage{
GenSamplingPointsBuffer(geom, ras, radius, cType)
}
\arguments{
\item{geom}{An sf geometry (single polygon)}

\item{ras}{A SpatRaster object}

\item{radius}{Buffer distance}

\item{cType}{"edge", "disk", or "donut"}
}
\value{
Matrix of (x, y) sampling coordinates on valid raster cells
}
\description{
Generate sampling points by buffering a polygon (sf-based)
}
\keyword{internal}
