% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.SpatialEffect}
\alias{print.SpatialEffect}
\title{Print method for SpatialEffect objects}
\usage{
\method{print}{SpatialEffect}(x, dVec.range = NULL, ...)
}
\arguments{
\item{x}{A SpatialEffect object}

\item{dVec.range}{Optional 2-element vector to filter displayed distances}

\item{...}{Ignored}
}
\description{
Print method for SpatialEffect objects
}
