% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle_average.R
\name{RimAvg}
\alias{RimAvg}
\title{Compute circle averages for all intervention nodes at a given distance}
\usage{
RimAvg(
  ras,
  Yobs = NULL,
  ras_Z = NULL,
  nz,
  Zdata,
  x_coord_Z,
  y_coord_Z,
  dUp,
  numpts = NULL,
  gridRes = NULL,
  evalpts = 10,
  only.unique = 0,
  dtype = "raster",
  dist.metric = "Euclidean",
  cType = "edge",
  dist_unit = NULL,
  n_threads = 1L
)
}
\arguments{
\item{ras}{SpatRaster or Krig object}

\item{Yobs}{Raster values vector (NULL for krig)}

\item{ras_Z}{sf object for polygon interventions (NULL for point)}

\item{nz}{Number of intervention nodes}

\item{Zdata}{Data frame of intervention data}

\item{x_coord_Z}{Name of x-coordinate column}

\item{y_coord_Z}{Name of y-coordinate column}

\item{dUp}{Distance value}

\item{numpts}{Number of sampling points}

\item{gridRes}{Raster resolution}

\item{evalpts}{Multiplier for numpts}

\item{only.unique}{Deduplicate grid cells?}

\item{dtype}{"raster" or "krig"}

\item{dist.metric}{"Euclidean" or "Geodesic"}

\item{cType}{"edge", "disk", or "donut"}

\item{dist_unit}{Step size for disk/donut}

\item{n_threads}{Number of threads for C++ computation}
}
\value{
List(Ybard, Ybard_sum, Ybard_len)
}
\description{
Compute circle averages for all intervention nodes at a given distance
}
\keyword{internal}
