% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialEffect.R
\name{.gen_perms}
\alias{.gen_perms}
\title{Generate permutation matrix}
\usage{
.gen_perms(Zup, blockvar = NULL, clustvar = NULL, maxiter = 1000)
}
\description{
Simple Bernoulli-style permutation matrix generator.
Falls back to the ri package if available and blockvar/clustvar are specified.
}
\keyword{internal}
