% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.SpatialEffect}
\alias{summary.SpatialEffect}
\title{Summary method for SpatialEffect objects}
\usage{
\method{summary}{SpatialEffect}(object, dVec.range = NULL, ...)
}
\arguments{
\item{object}{A SpatialEffect object}

\item{dVec.range}{Optional 2-element vector to filter displayed distances}

\item{...}{Ignored}
}
\description{
Summary method for SpatialEffect objects
}
